AAH_FLUDSCommonData is in 1 .h/.cc file
